+++ backend/core/src/units-summary/dto/units-summary.dto.ts
@@ -1,21 +1,20 @@
-import { OmitType } from "@nestjs/swagger"
-import { Expose, Type } from "class-transformer"
-import { IsString, IsUUID, ValidateNested } from "class-validator"
+import { ApiHideProperty, OmitType } from "@nestjs/swagger"
+import { Exclude, Expose, Type } from "class-transformer"
+import { IsOptional, IsUUID } from "class-validator"
 import { IdDto } from "../../shared/dto/id.dto"
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { UnitsSummary } from "../entities/units-summary.entity"
 
 export class UnitsSummaryDto extends OmitType(UnitsSummary, ["listing"] as const) {
-  @Expose()
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  listing: IdDto
+  @Exclude()
+  @ApiHideProperty()
+  listing
 }
 
 export class UnitsSummaryCreateDto extends OmitType(UnitsSummaryDto, ["id"] as const) {}
 export class UnitsSummaryUpdateDto extends OmitType(UnitsSummaryCreateDto, [] as const) {
   @Expose()
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsUUID()
-  id: string
+  id?: string
 }