+++ sites/partners/src/listings/PaperListingForm/index.tsx
@@ -31,6 +31,7 @@ import {
   ListingEventType,
   ListingEventCreate,
   Preference,
+  UnitsSummary,
   ListingReviewOrder,
 } from "@bloom-housing/backend-core/types"
 import { YesNoAnswer } from "../../applications/PaperApplicationForm/FormTypes"
@@ -43,7 +44,13 @@ import AdditionalEligibility from "./sections/AdditionalEligibility"
 import LeasingAgent from "./sections/LeasingAgent"
 import AdditionalFees from "./sections/AdditionalFees"
 import Units from "./sections/Units"
-import { stringToBoolean, stringToNumber, createDate, createTime } from "../../../lib/helpers"
+import {
+  stringToBoolean,
+  stringToNumberOrOne,
+  createDate,
+  createTime,
+  toNumberOrNull,
+} from "../../../lib/helpers"
 import BuildingDetails from "./sections/BuildingDetails"
 import ListingIntro from "./sections/ListingIntro"
 import ListingPhoto from "./sections/ListingPhoto"
@@ -212,6 +219,10 @@ export type TempUnit = Unit & {
   tempId?: number
 }
 
+export type TempUnitsSummary = UnitsSummary & {
+  tempId?: number
+}
+
 export type TempEvent = ListingEvent & {
   tempId?: string
 }
@@ -220,18 +231,13 @@ const formatFormData = (
   data: FormListing,
   units: TempUnit[],
   openHouseEvents: TempEvent[],
+  summaries: TempUnitsSummary[],
   preferences: Preference[],
   saveLatLong: LatitudeLongitude,
   customPinPositionChosen: boolean
 ) => {
   const showWaitlistNumber =
     data.waitlistOpenQuestion === YesNoAnswer.Yes && data.waitlistSizeQuestion === YesNoAnswer.Yes
-
-  const applicationDueDateFormatted = createDate(data.applicationDueDateField)
-  const applicationDueTimeFormatted = createTime(
-    applicationDueDateFormatted,
-    data.applicationDueTimeField
-  )
 
   units.forEach((unit) => {
     switch (unit.unitType?.name) {
@@ -251,16 +257,44 @@ const formatFormData = (
         unit.numBedrooms = null
     }
 
-    unit.floor = stringToNumber(unit.floor)
-    unit.maxOccupancy = stringToNumber(unit.maxOccupancy)
-    unit.minOccupancy = stringToNumber(unit.minOccupancy)
-    unit.numBathrooms = stringToNumber(unit.numBathrooms)
+    unit.floor = stringToNumberOrOne(unit.floor)
+    unit.maxOccupancy = stringToNumberOrOne(unit.maxOccupancy)
+    unit.minOccupancy = stringToNumberOrOne(unit.minOccupancy)
+    unit.numBathrooms = stringToNumberOrOne(unit.numBathrooms)
 
     if (!unit.sqFeet) {
       delete unit.sqFeet
     }
 
     delete unit.tempId
+  })
+  summaries.forEach((summary) => {
+    summary.minOccupancy = toNumberOrNull(summary.minOccupancy)
+    summary.maxOccupancy = toNumberOrNull(summary.maxOccupancy)
+    summary.floorMin = toNumberOrNull(summary.floorMin)
+    summary.floorMax = toNumberOrNull(summary.floorMax)
+    summary.totalAvailable = toNumberOrNull(summary.totalAvailable)
+    summary.totalCount = toNumberOrNull(summary.totalCount)
+    summary.monthlyRentMin = toNumberOrNull(summary.monthlyRentMin)
+    summary.monthlyRentMax = toNumberOrNull(summary.monthlyRentMax)
+
+    if (!summary.sqFeetMin) {
+      delete summary.sqFeetMin
+    }
+    if (!summary.sqFeetMax) {
+      delete summary.sqFeetMax
+    }
+    if (!summary.minimumIncomeMin) {
+      delete summary.minimumIncomeMin
+    }
+    if (!summary.monthlyRentAsPercentOfIncome) {
+      delete summary.monthlyRentAsPercentOfIncome
+    }
+
+    if (summary.id === undefined) {
+      delete summary.id
+    }
+    delete summary.tempId
   })
 
   const events: ListingEventCreate[] = data.events.filter(
@@ -295,7 +329,6 @@ const formatFormData = (
 
   return {
     ...data,
-    applicationDueTime: applicationDueTimeFormatted,
     disableUnitsAccordion: stringToBoolean(data.disableUnitsAccordion),
     units: units,
     preferences: preferences,
@@ -306,7 +339,6 @@ const formatFormData = (
     },
     customMapPin: customPinPositionChosen,
     isWaitlistOpen: data.waitlistOpenQuestion === YesNoAnswer.Yes,
-    applicationDueDate: applicationDueDateFormatted,
     yearBuilt: data.yearBuilt ? Number(data.yearBuilt) : null,
     waitlistCurrentSize:
       data.waitlistCurrentSize && showWaitlistNumber ? Number(data.waitlistCurrentSize) : null,
@@ -341,6 +373,7 @@ const formatFormData = (
       : null,
     events,
     reservedCommunityType: data.reservedCommunityType.id ? data.reservedCommunityType : null,
+    unitsSummary: summaries,
     reviewOrderType:
       data.reviewOrderQuestion === "reviewOrderLottery"
         ? ListingReviewOrder.lottery
@@ -365,6 +398,7 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
   const [status, setStatus] = useState<ListingStatus>(null)
   const [submitData, setSubmitData] = useState<SubmitData>({ ready: false, data: defaultValues })
   const [units, setUnits] = useState<TempUnit[]>([])
+  const [unitsSummaries, setUnitsSummaries] = useState<TempUnitsSummary[]>([])
   const [openHouseEvents, setOpenHouseEvents] = useState<TempEvent[]>([])
   const [preferences, setPreferences] = useState<Preference[]>(listing?.preferences ?? [])
   const [latLong, setLatLong] = useState<LatitudeLongitude>({
@@ -414,6 +448,15 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
 
       setOpenHouseEvents(events)
     }
+
+    // Use a temp id to track each summary within the form table (prior to submission).
+    if (listing?.unitsSummary) {
+      const tempSummaries = listing.unitsSummary.map((summary, i) => ({
+        ...summary,
+        tempId: i + 1,
+      }))
+      setUnitsSummaries(tempSummaries)
+    }
   }, [listing, setUnits, setOpenHouseEvents])
 
   // eslint-disable-next-line @typescript-eslint/unbound-method
@@ -442,6 +485,7 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
           data,
           units,
           openHouseEvents,
+          unitsSummaries,
           orderedPreferences,
           latLong,
           customMapPositionChosen
@@ -469,6 +513,7 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
     },
     [
       units,
+      unitsSummaries,
       openHouseEvents,
       editMode,
       listingsService,
@@ -557,7 +602,8 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
                           <Units
                             units={units}
                             setUnits={setUnits}
-                            unitsSummary={listing?.unitsSummarized}
+                            unitsSummaries={unitsSummaries}
+                            setSummaries={setUnitsSummaries}
                             disableUnitsAccordion={listing?.disableUnitsAccordion}
                           />
                           <Preferences preferences={preferences} setPreferences={setPreferences} />